fix return type ``::Int64``

+++ src/fun_fs/elastoplast.jl
@@ -96,5 +96,5 @@ end
             mpD.σ[:,p] .= mpD.τ[:,p]./mpD.J[p]
         end
     end
-    return ηmax
+    return ηmax::Int64
 end
